requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
platformdirs==3.5.3
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
Werkzeug==2.3.6
